feat!: Drop edx-name-affirmation as a dependency.
This dependency is specific to edx.org and should not be in the default
version of the edx-platform.  There is related code in edx-platform but
circuit breakers should keep that from running for now until we can
clean it up.

BREAKING CHANGE: If you are relying on the edx-name-affirmation app
working, you should install it yourself before running the platform.